Refactor dashboard codebase (#124)
