+++ dashboard/src/App.tsx
@@ -3,9 +3,9 @@ import { ThemeProvider, createTheme } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
 import { Box, Container, Link, Typography, Button } from '@mui/material';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Dashboard from './components/Dashboard';
-import ShareDetails from './components/ShareDetails';
-import MinedSharesExplorer from './components/MinerDashboard';
+import Dashboard from './components/Dashboard/Dashboard';
+import ShareDetails from './components/ShareDetails/ShareDetails';
+import MinedSharesExplorer from './components/MinerDashboard/MinerDashboard';
 
 // Create a dark theme
 const theme = createTheme({
