+++ dashboard/src/components/Dashboard/PoolHashrateChart.tsx
@@ -7,8 +7,8 @@ import {
   Alert,
 } from '@mui/material';
 import * as d3 from 'd3';
-import colors from '../theme/colors';
-import Card from './common/Card';
+import colors from '../../theme/colors';
+import Card from '../common/Card';
 
 // Mock data for the hashrate over time
 const mockHashrateData = [
