+++ dashboard/src/components/MinerDashboard/Filters.tsx
