+++ dashboard/src/components/MinerDashboard/MinerChart.tsx
