+++ dashboard/src/components/MinerDashboard/MinerTable.tsx
