+++ dashboard/src/components/MinerDashboard/Particles.tsx
