+++ dashboard/src/components/common/KPICard.tsx
